refactor: customized errors

+++ src/common/base/base.service.ts
@@ -1,6 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { DeepPartial, Repository } from 'typeorm';
 import { BaseEntity } from './base.entity';
+import { EntityNotFoundException } from '../exceptions';
 
 @Injectable()
 export abstract class BaseService<T extends BaseEntity> {
@@ -26,7 +27,7 @@ export abstract class BaseService<T extends BaseEntity> {
     });
 
     if (!entity) {
-      throw new NotFoundException(`${this.entityName} with id ${id} not found`);
+      throw new EntityNotFoundException(this.entityName, id);
     }
 
     return entity;
@@ -35,6 +36,10 @@ export abstract class BaseService<T extends BaseEntity> {
   async update(id: number, data: DeepPartial<T>): Promise<T> {
     const entity = await this.findById(id);
 
+    if (!entity) {
+      throw new EntityNotFoundException(this.entityName, id);
+    }
+
     this.repository.merge(entity, data);
 
     return this.repository.save(entity);
@@ -42,6 +47,10 @@ export abstract class BaseService<T extends BaseEntity> {
 
   async delete(id: number): Promise<void> {
     const entity = await this.findById(id);
+
+    if (!entity) {
+      throw new EntityNotFoundException(this.entityName, id);
+    }
 
     await this.repository.remove(entity);
   }
